Move FeatureFlag to source, add ExternBody node and fix structural visitors
