+++ effekt/shared/src/main/scala/effekt/Namer.scala
@@ -294,7 +294,7 @@ object Namer extends Phase[Parsed, NameResolved] {
         sym.tparams.foreach { p => Context.bind(p) }
         Context.bindValues(sym.vparams)
         Context.bindBlocks(sym.bparams)
-        bodies.foreach{ (ff, body) => body.args.foreach(resolveGeneric) }
+        bodies.foreach{ case source.ExternBody(ff, body) => body.args.foreach(resolveGeneric) }
       }
 
     case source.InterfaceDef(id, tparams, operations, isEffect) =>
