+++ effekt/shared/src/main/scala/effekt/core/Transformer.scala
@@ -103,7 +103,7 @@ object Transformer extends Phase[Typechecked, CoreTransformed] {
       val sym@ExternFunction(name, tps, _, _, ret, effects, capt, _) = f.symbol
       assert(effects.isEmpty)
       val cps = bps.map(b => b.symbol.capture)
-      val tBodies = bodies.map { (ff, body) =>
+      val tBodies = bodies.map { case source.ExternBody(ff, body) =>
         val args = body.args.map(transformAsExpr).map {
           case p: Pure => p: Pure
           case _ => Context.abort("Spliced arguments need to be pure expressions.")
