+++ effekt/shared/src/main/scala/effekt/core/Tree.scala
@@ -1,7 +1,7 @@
 package effekt
 package core
 
-import effekt.FeatureFlag
+import effekt.source.FeatureFlag
 import effekt.util.Structural
 import effekt.util.messages.INTERNAL_ERROR
